agentshield/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentshield"
version = "0.1.0"
description = "Security firewall for AI agents — block prompt injection, data exfiltration, and unauthorized tool access."
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
authors = [
    { name = "Your Name", email = "you@example.com" },
]
keywords = [
    "ai", "security", "agents", "llm", "mcp",
    "prompt-injection", "firewall", "langchain",
    "model-context-protocol", "guardrails",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
presidio = [
    "presidio-analyzer>=2.2",
    "presidio-anonymizer>=2.2",
]
langchain = [
    "langchain-core>=0.2",
]
all = [
    "agentshield[presidio,langchain]",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "mypy>=1.10",
    "pre-commit>=3.7",
]

[project.scripts]
agentshield = "agentshield.cli.main:cli"

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/agentshield"
Documentation = "https://github.com/YOUR_USERNAME/agentshield/tree/main/docs"
Repository = "https://github.com/YOUR_USERNAME/agentshield"
Issues = "https://github.com/YOUR_USERNAME/agentshield/issues"

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = ["S101"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true

[tool.hatch.build.targets.wheel]
packages = ["agentshield"]